use correct size for sigs

+++ include/makesig.h
@@ -13,6 +13,6 @@
 #ifndef _MAKESIG_H
 #define _MAKESIG_H
 
-#define MAKE_SIG(a,b,c,d) (((long)(a)<<24)+((long)(b)<<16)+((c)<<8)+(d))
+#define MAKE_SIG(a,b,c,d) (((uint32_t)(a)<<24)+((uint32_t)(b)<<16)+((c)<<8)+(d))
 
 #endif
